+++ Runtime/Components/Editor/SpriteAnimatorEditor.cs
@@ -1,0 +1,12 @@
+#if UNITY_EDITOR
+using UnityEditor;
+
+namespace NekoLib.Components
+{
+    [CustomEditor(typeof(SpriteAnimator))]
+    public class SpriteAnimatorEditor : BaseSpriteAnimatorEditor
+    {
+
+    }
+}
+#endif